backend/requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0

# AI/ML libraries (updated versions)
torch>=2.6.0
torchvision>=0.19.0
transformers>=4.40.0
accelerate>=0.30.0

# Image processing
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# Additional utilities
requests==2.31.0
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
# For SQLite (development)
# For MySQL (production) - uncomment if needed:
# mysql-connector-python==8.2.0

# Testing (for later tasks)
pytest==7.4.3
pytest-cov==4.1.0
